New architecture and some adjustments. (#26)
* Adding my Tests - VM

* Test de Valentin

* Database Connection Added.

I added the Database connection so we can store objects in a database.

* POST Method & Database.

I added the feature to save the created todo into the database.

* Post Method & Database Connection.

I added the feature to save the created todo into the database.

* GET List of Todos Method.

Method adapted to obtain a list of Todos that are stored in the database.

* GET Specific Todo Method.

https://trello.com/c/9aDVyEYO/22-crear-un-endpoint-para-listar-un-todo-espec%C3%ADfico
Method adapted to obtain a specific Todo that is stored in the database.

* PUT Method Added.

https://trello.com/c/UKryHi0C/10-crear-un-endpoint-para-editar-un-todo
New method added to modify a specific Todo that is stored in the database.

* DELETE Method Added.

https://trello.com/c/EjojPGnO/9-crear-un-endpoint-para-eliminar-un-todo
103 / 5,000
Resultados de traducción
Traducción
Added new method to delete a specific Todo that is stored in the database.

* Architecture update and other features.

Task: https://trello.com/c/5gLSPx7q/25-adaptar-arquitectura-a-los-archivos-existentes
Se ha implementado una reestructuración de la arquitectura del proyecto:
> backend
	- __init__.py
	- .gitignore
	- main.py
	- pyproject.toml
	- requirements.txt
	- selector.py
	- services.py
	> todo
  		> adapter
			- mongo.py
		> port
			- database.py
			- models.py
			- todo.py

* Filtering Added.

Added filtered_by() method as a option when using get_todos_list().
So you can get a filtered list of todos as completed or uncompleted, or you can get all todos (completed and uncompleted)

* Little Upgrade to filtered_by()

Added the feature to make lower case the words "completed" and "uncompleted" when using filtered_by() method.

* Errors Fixed.

Fix all errors that reviewers remarked.

* Missing Update in Main Module.

Updated the type of data that return update_todo and get_todos_list function.

* New Changes [Before Fixing Main Branch Conflicts]

Fixing problems in mongo.py and database.py modules.

* Validations Added.

I added some validations to the code. Probably there are many validations missing but when I detect them, they will be added soon.

* Fixed Implemented Errors.

I have corrected the errors that were in the form of raising errors previously.